:recycle:refactor: refactor one line review dialog 상수, 함수, 컴포넌트, 스타일 분리 constants 폴더 생성 & 상수 관리 Related to: #78
