+++ src/components/FeedPage/WriteDialog.tsx
@@ -2,7 +2,7 @@ import HybridDialog from '@components/commons/HybridDialog';
 import { Button, Stack } from '@mui/material';
 import { useState } from 'react';
 import PostingDialog from './PostingDialog';
-import OneLineReviewDialog from './OneLineReviewDialog';
+import OneLineReviewDialog from './OneLineReviewDialog/OneLineReviewDialog';
 import PostAddIcon from '@mui/icons-material/PostAdd';
 import BorderColorIcon from '@mui/icons-material/BorderColor';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
