+++ src/constants/dialog.ts
@@ -1,0 +1,4 @@
+export const REVIEW_DIALOG = {
+  ONE_LINE_REVIEW_MAX_LENGTH: 50,
+  DEFAULT_STAR_RATING: 0,
+} as const;
